backend/requirements.txt
# This file is auto-generated by uv pip compile pyproject.toml
# To update, run: uv pip compile pyproject.toml
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic[email]>=2.5.0
pydantic-settings>=2.1.0
sqlalchemy>=2.0.23
alembic>=1.12.1
asyncpg>=0.29.0
psycopg2-binary>=2.9.7
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
httpx>=0.25.2
celery[redis]>=5.3.4
redis>=5.0.1
supabase>=2.0.2
python-dotenv>=1.0.0
pytest>=7.4.3
pytest-asyncio>=0.21.1
httpx>=0.25.2
aiosqlite>=0.19.0